# Conflicts:
#	.gitignore
#	README.md
#	clean.sh
#	copy.sh
#	deploy.sh
#	docker-compose.yml
#	docker/mysql/db/im_core.sql
#	im-domain/Dockerfile
#	im-domain/src/main/resources/application-docker.yml
#	im-domain/src/main/resources/application.yml
#	im-message-store/Dockerfile
#	im-message-store/src/main/resources/application-docker.yml
#	im-message-store/src/main/resources/application.yml
#	im-tcp/Dockerfile
#	im-tcp/im-tcp.iml
#	im-tcp/pom.xml
#	im-tcp/src/main/resources/WebSocket.html
#	im-tcp/src/main/resources/config-docker.yml
#	im-tcp/src/main/resources/static/websocket.html
#	logback-spring.xml
#	open_port.sh
#	pom.xml

+++ clean.sh
@@ -5,6 +5,12 @@ echo "::: Welcome to IM-WhaleShark :::"
 
 # 删除 build 文件夹
 echo "开始清理 build 文件夹 .."
-
-rm -rf docker/build/
+if [ -d "docker/build" ]; then
+  DELETED_FILES=$(find docker/build -type f -o -type d)
+  rm -rf docker/build/
+  echo "以下文件夹和文件已被删除:"
+  echo "$DELETED_FILES" | sed 's/ /\n/g'
+else
+  echo "build 文件夹不存在，无需清理。"
+fi
 echo "清理完成"
